requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchaudio==2.5.1

annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
openai==1.55.3
packaging==24.2
praat-parselmouth==0.4.5
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
soundfile==0.12.1
sympy==1.13.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3